+++ pysatNASA/instruments/timed_saber.py
@@ -104,8 +104,9 @@ clean = mm_nasa.clean_warn
 # Use the default CDAWeb and pysat methods
 
 # Set the list_files routine
-fname = ''.join(('timed_l2av207_saber_{year:04d}{month:02d}{day:02d}',
-                 '{hour:02d}{minute:02d}_v{version:02d}.cdf'))
+fname = ''.join(('timed_l2a_saber_{year:04d}{month:02d}{day:02d}',
+                 '{hour:02d}{minute:02d}_v{version:02d}-{revision:02d}-',
+                 '{cycle:02d}.cdf'))
 supported_tags = {'': {'': fname}}
 list_files = functools.partial(mm_gen.list_files,
                                supported_tags=supported_tags)
@@ -114,7 +115,7 @@ list_files = functools.partial(mm_gen.list_files,
 load = cdw.load
 
 # Set the download routine
-basic_tag = {'remote_dir': ''.join(('/pub/data/timed/saber/level2a_v2_07_cdf',
+basic_tag = {'remote_dir': ''.join(('/pub/data/timed/saber/level2a_cdf',
                                     '/{year:4d}/{month:02d}/')),
              'fname': fname}
 download_tags = {'': {'': basic_tag}}